backend/requirements.txt
aiosqlite==0.22.1
fastapi==0.135.3
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-oauthlib==1.3.1
google-generativeai==0.8.6
langgraph==1.1.6
mcp==1.27.0
networkx==3.4.2
pydantic==2.12.5
python-dotenv==1.2.2
SQLAlchemy==2.0.49
sse-starlette==3.3.4
uvicorn==0.44.0
